Partial start class
