+++ ib.bat
@@ -1,0 +1,2 @@
+@echo off
+dotnet build
