+++ src/Start.cs
@@ -24,7 +24,7 @@ namespace jammer
         next
     }
 
-    public class Start
+    public partial class Start
     {
         //NOTE(ra) Starting state to playing. 
         // public static MainStates state = MainStates.idle;
@@ -135,59 +135,6 @@ namespace jammer
                         Play.NextSong();
                         Play.PlaySong(Utils.songs, Utils.currentSongIndex);
                         state = MainStates.play;
-                        break;
-
-
-                }
-            }
-        }
-
-        public static void CheckKeyboard()
-        {
-            if (Console.KeyAvailable)
-            {
-                var key = Console.ReadKey(true).Key;
-                switch (key)
-                {
-                    case ConsoleKey.Spacebar:
-                        if (Raylib.IsMusicReady(Utils.currentMusic) && !Raylib.IsMusicStreamPlaying(Utils.currentMusic))
-                        {
-                            Play.PlaySong();
-                            state = MainStates.playing;
-                            drawOnce = true;
-                        }
-                        else if (Raylib.IsMusicStreamPlaying(Utils.currentMusic))
-                        {
-                            Console.WriteLine("Paused");
-                            state = MainStates.pause;
-                            drawOnce = true;
-                        }
-                        else
-                        //NOTE(ra) Resumed is not called at all. PlaySong resumes after pause.
-                        {
-                            Console.WriteLine("Resumed");
-                            Play.ResumeSong();
-                        }
-                        break;
-                    case ConsoleKey.F12:
-                        Console.WriteLine("CurrentState: " + state);
-                        break;
-                    case ConsoleKey.Q:
-                        Console.WriteLine("Quit");
-                        AnsiConsole.Clear();
-                        Environment.Exit(0);
-                        break;
-
-                    case ConsoleKey.Escape:
-                        AnsiConsole.Clear();
-                        Console.WriteLine("Quit");
-                        Environment.Exit(0);
-                        break;
-                    case ConsoleKey.RightArrow:
-                        state = MainStates.next;
-                        break;
-                    case ConsoleKey.LeftArrow:
-                        Play.PrevSong();
                         break;
                 }
             }
